Numeric.Units.Dimensional

Description

Please refer to the literate Haskell code for documentation of both API and implementation.

Documentation

newtype Dimensional v d a

Constructors

Dimensional a

data DUnit

data DQuantity

type Unit = Dimensional DUnit

type Quantity = Dimensional DQuantity

(*~) :: Num a => a -> Unit d a -> Quantity d a infixl 7

(/~) :: Fractional a => Quantity d a -> Unit d a -> a infixl 7

data Dim l m t i th n j

type DOne = Dim Zero Zero Zero Zero Zero Zero Zero

type DLength = Dim Pos1 Zero Zero Zero Zero Zero Zero

type DMass = Dim Zero Pos1 Zero Zero Zero Zero Zero

type DTime = Dim Zero Zero Pos1 Zero Zero Zero Zero

type DElectricCurrent = Dim Zero Zero Zero Pos1 Zero Zero Zero

type DThermodynamicTemperature = Dim Zero Zero Zero Zero Pos1 Zero Zero

type DAmountOfSubstance = Dim Zero Zero Zero Zero Zero Pos1 Zero

type DLuminousIntensity = Dim Zero Zero Zero Zero Zero Zero Pos1

type Dimensionless = Quantity DOne

type Length = Quantity DLength

type Mass = Quantity DMass

type Time = Quantity DTime

type ElectricCurrent = Quantity DElectricCurrent

type ThermodynamicTemperature = Quantity DThermodynamicTemperature

type AmountOfSubstance = Quantity DAmountOfSubstance

type LuminousIntensity = Quantity DLuminousIntensity

class Mul d d' d'' | d d' -> d''

class Div d d' d'' | d d' -> d''

class NumType x => Pow d x d' | d x -> d'

class NonZero x => Root d x d' | d x -> d'

(*) :: (Num a, Mul d d' d'') => Dimensional v d a -> Dimensional v d' a -> Dimensional v d'' a infixl 7

(/) :: (Fractional a, Div d d' d'') => Dimensional v d a -> Dimensional v d' a -> Dimensional v d'' a infixl 7

(^) :: (Fractional a, Pow d n d') => Dimensional v d a -> n -> Dimensional v d' a infixr 8

(^+) :: (Num a, PosType n, Pow d n d') => Dimensional v d a -> n -> Dimensional v d' a infixr 8

negate :: Num a => Quantity d a -> Quantity d a

(+) :: Num a => Quantity d a -> Quantity d a -> Quantity d a infixl 6

(-) :: Num a => Quantity d a -> Quantity d a -> Quantity d a infixl 6

abs :: Num a => Quantity d a -> Quantity d a

nroot :: (Floating a, Root d n d') => n -> Dimensional v d a -> Dimensional v d' a

sqrt :: (Floating a, Root d Pos2 d') => Dimensional v d a -> Dimensional v d' a

cbrt :: (Floating a, Root d Pos3 d') => Dimensional v d a -> Dimensional v d' a

(^/) :: (Floating a, Root d n d') => Dimensional v d a -> n -> Dimensional v d' a infixr 8

(*~~) :: (Functor f, Num a) => f a -> Unit d a -> f (Quantity d a) infixl 7

(/~~) :: (Functor f, Fractional a) => f (Quantity d a) -> Unit d a -> f a infixl 7

sum :: forall d a. Num a => [Quantity d a] -> Quantity d a

dimensionlessLength :: Num a => [Dimensional v d a] -> Dimensionless a

exp :: Floating a => Dimensionless a -> Dimensionless a

atanh :: Floating a => Dimensionless a -> Dimensionless a

acosh :: Floating a => Dimensionless a -> Dimensionless a

asinh :: Floating a => Dimensionless a -> Dimensionless a

tanh :: Floating a => Dimensionless a -> Dimensionless a

cosh :: Floating a => Dimensionless a -> Dimensionless a

sinh :: Floating a => Dimensionless a -> Dimensionless a

atan :: Floating a => Dimensionless a -> Dimensionless a

acos :: Floating a => Dimensionless a -> Dimensionless a

asin :: Floating a => Dimensionless a -> Dimensionless a

tan :: Floating a => Dimensionless a -> Dimensionless a

cos :: Floating a => Dimensionless a -> Dimensionless a

sin :: Floating a => Dimensionless a -> Dimensionless a

log :: Floating a => Dimensionless a -> Dimensionless a

(**) :: Floating a => Dimensionless a -> Dimensionless a -> Dimensionless a infixr 8

atan2 :: RealFloat a => Quantity d a -> Quantity d a -> Dimensionless a

one :: Num a => Unit DOne a

_0 :: Num a => Quantity d a

_1 :: Num a => Dimensionless a

_9 :: Num a => Dimensionless a

_8 :: Num a => Dimensionless a

_7 :: Num a => Dimensionless a

_6 :: Num a => Dimensionless a

_5 :: Num a => Dimensionless a

_4 :: Num a => Dimensionless a

_3 :: Num a => Dimensionless a

_2 :: Num a => Dimensionless a

pi :: Floating a => Dimensionless a

tau :: Floating a => Dimensionless a

dimUnit :: NumType n => String -> n -> Maybe String

prefix :: Num a => a -> Unit d a -> Unit d a