Numeric.Units.Dimensional.CGS

Description

Please refer to the literate Haskell code for documentation of both API and implementation.

Documentation

data CGSDim lh mh t

type DLength = CGSDim Pos2 Zero Zero

type DMass = CGSDim Zero Pos2 Zero

type DTime = CGSDim Zero Zero Pos1

type DElectricCurrent = CGSDim Pos3 Pos1 Neg2

type DCharge = CGSDim Pos3 Pos1 Neg1

meter :: Num a => Unit DLength a

gram :: Num a => Unit DMass a

second :: Num a => Unit DTime a

ampere :: Floating a => Unit DElectricCurrent a

franklin :: Floating a => Unit DCharge a

unit_SI :: Num a => Quantity (Dim l m t i th n j) a -> Unit (Dim l m t i th n j) a

unit_CGS :: forall a l m t i l2 m2 il it l' m' t'. (Floating a, Mul Zero l Zero, Mul Pos2 l l2, Mul Zero m Zero, Mul Pos2 m m2, Mul Zero t Zero, Mul Pos1 t t, Sum l2 Zero l2, Sum Zero m2 m2, Sum m2 Zero m2, Sum Zero t t, Mul Pos3 i il, Mul Pos1 i i, Mul Neg2 i it, Sum l2 il l', Sum m2 i m', Sum t it t') => Quantity (Dim l m t i Zero Zero Zero) a -> Unit (CGSDim l' m' t') a

fromSI :: (Mul Neg2 i it, Mul Pos3 i il, Mul Pos2 m m2, Mul Pos2 l l2, Mul Pos1 i i, Mul Pos1 t t, Mul Zero t Zero, Mul Zero m Zero, Mul Zero l Zero, Sub t' it t, Sub t' t it, Sub m' m2 i, Sub m' i m2, Sub l' il l2, Sub l' l2 il, Sub m2 m2 Zero, Sub l2 l2 Zero, Sub t t Zero, Add m2 Zero m2, Add m2 i m', Add l2 Zero l2, Add l2 il l', Add t it t', Floating a) => Quantity (Dim l m t i Zero Zero Zero) a -> Quantity (CGSDim l' m' t') a

toSI :: forall a l m t i l2 m2 il it l' m' t'. (Floating a, Mul Zero l Zero, Mul Pos2 l l2, Mul Zero m Zero, Mul Pos2 m m2, Mul Zero t Zero, Mul Pos1 t t, Sum l2 Zero l2, Sum Zero m2 m2, Sum m2 Zero m2, Sum Zero t t, Mul Pos3 i il, Mul Pos1 i i, Mul Neg2 i it, Sum l2 il l', Sum m2 i m', Sum t it t') => Quantity (CGSDim l' m' t') a -> Quantity (Dim l m t i Zero Zero Zero) a

q_si :: Quantity DElectricCharge Double

r_si :: Quantity DLength Double

f_si :: Dimensional DQuantity (Dim (Pos Zero) (Pos Zero) (Neg (Neg Zero)) Zero Zero Zero Zero) Double

q_cgs :: Quantity (CGSDim (Pos (Pos (Pos Zero))) Pos1 (Neg Zero)) Double

r_cgs :: Quantity (CGSDim (Pos (Pos Zero)) Zero Zero) Double

f_cgs :: Dimensional DQuantity (CGSDim (Pos (Pos Zero)) (Pos (Pos Zero)) (Neg (Neg Zero))) Double

f_si' :: Force Double

c :: Quantity DCapacitance Double

c_cgs :: Quantity (CGSDim (Pos (Pos Zero)) Zero Zero) Double

c' :: Capacitance Double

c'' :: Length Double