Numeric.Units.Dimensional.Dimensions

Description

Provides both term-level and type-level representations for physical dimensions in a single import for convenience.

Presuming that users intend to work primarily with type level dimensions, this module hides arithmetic operators over term level dimensions and aliases for the base term-level dimensions to avoid namespace pollution. These features are available directly from Numeric.Units.Dimensional.Dimensions.TermLevel if desired.

Documentation

class HasDynamicDimension a => HasDimension a where

Dimensional values inhabit this class, which allows access to a term-level representation of their dimension.

Methods

dimension :: a -> Dimension'

Obtains a term-level representation of a value's dimension.

class HasDynamicDimension a where

Dimensional values, or those that are only possibly dimensional, inhabit this class, which allows access to a term-level representation of their dimension.

Minimal complete definition

Nothing

Methods

dynamicDimension :: a -> DynamicDimension

Gets the 'DynamicDimension of a dynamic dimensional value, which may be NoDimension if it does not represent a dimensional value of any Dimension.

A default implementation is available for types that are also in the HasDimension typeclass.

default dynamicDimension :: HasDimension a => a -> DynamicDimension

data DynamicDimension

The dimension of a dynamic value, which may not have any dimension at all.

Constructors

NoDimension
The value has no valid dimension.

SomeDimension Dimension'
The value has the given dimension.

AnyDimension
The value may be interpreted as having any dimension.

data Dimension'

A physical dimension, encoded as 7 integers, representing a factorization of the dimension into the 7 SI base dimensions. By convention they are stored in the same order as in the Dimension data kind.

Constructors

Dim' !Int !Int !Int !Int !Int !Int !Int

matchDimensions :: DynamicDimension -> DynamicDimension -> DynamicDimension

Combines two DynamicDimensions, determining the DynamicDimension of a quantity that must match both inputs.

This is the lattice meet operation for DynamicDimension.

isCompatibleWith :: HasDynamicDimension a => a -> Dimension' -> Bool

Determines if a value that has a DynamicDimension is compatible with a specified Dimension'.

hasSomeDimension :: HasDynamicDimension a => a -> Bool

Determines if a value that has a DynamicDimension in fact has any valid dimension at all.

dOne :: Dimension'

The dimension of dimensionless values.

asList :: Dimension' -> [Int]

Converts a dimension to a list of 7 integers, representing the exponent associated with each of the 7 SI base dimensions in the standard order.

module Numeric.Units.Dimensional.Dimensions.TypeLevel