Numeric.Units.Dimensional.Dimensions.TermLevel

Description

This module defines physical dimensions expressed in terms of the SI base dimensions, including arithmetic.

Type

data Dimension'

A physical dimension, encoded as 7 integers, representing a factorization of the dimension into the 7 SI base dimensions. By convention they are stored in the same order as in the Dimension data kind.

Constructors

Dim' !Int !Int !Int !Int !Int !Int !Int

Access to Dimension of Dimensional Values

class HasDynamicDimension a => HasDimension a where

Dimensional values inhabit this class, which allows access to a term-level representation of their dimension.

Methods

dimension :: a -> Dimension'

Obtains a term-level representation of a value's dimension.

class HasDynamicDimension a where

Dimensional values, or those that are only possibly dimensional, inhabit this class, which allows access to a term-level representation of their dimension.

Minimal complete definition

Nothing

Methods

dynamicDimension :: a -> DynamicDimension

Gets the 'DynamicDimension of a dynamic dimensional value, which may be NoDimension if it does not represent a dimensional value of any Dimension.

A default implementation is available for types that are also in the HasDimension typeclass.

default dynamicDimension :: HasDimension a => a -> DynamicDimension

data DynamicDimension

The dimension of a dynamic value, which may not have any dimension at all.

Constructors

NoDimension
The value has no valid dimension.

SomeDimension Dimension'
The value has the given dimension.

AnyDimension
The value may be interpreted as having any dimension.

Dimension Arithmetic

(*) :: Dimension' -> Dimension' -> Dimension' infixl 7

Forms the product of two dimensions.

(/) :: Dimension' -> Dimension' -> Dimension' infixl 7

Forms the quotient of two dimensions.

(^) :: Dimension' -> Int -> Dimension' infixr 8

Raises a dimension to an integer power.

recip :: Dimension' -> Dimension'

Forms the reciprocal of a dimension.

nroot :: Int -> Dimension' -> Maybe Dimension'

Takes the nth root of a dimension, if it exists.

n must not be zero.

nroot (negate n) d == nroot n (recip d)

sqrt :: Dimension' -> Maybe Dimension'

Takes the square root of a dimension, if it exists.

sqrt d == nroot 2 d

cbrt :: Dimension' -> Maybe Dimension'

Takes the cube root of a dimension, if it exists.

cbrt d == nroot 3 d

Synonyms for Base Dimensions

dOne :: Dimension'

The dimension of dimensionless values.

dLength :: Dimension'

dMass :: Dimension'

dTime :: Dimension'

dElectricCurrent :: Dimension'

dThermodynamicTemperature :: Dimension'

dAmountOfSubstance :: Dimension'

dLuminousIntensity :: Dimension'

Deconstruction

asList :: Dimension' -> [Int]

Converts a dimension to a list of 7 integers, representing the exponent associated with each of the 7 SI base dimensions in the standard order.

Examining Dynamic Dimensions

matchDimensions :: DynamicDimension -> DynamicDimension -> DynamicDimension

Combines two DynamicDimensions, determining the DynamicDimension of a quantity that must match both inputs.

This is the lattice meet operation for DynamicDimension.

isCompatibleWith :: HasDynamicDimension a => a -> Dimension' -> Bool

Determines if a value that has a DynamicDimension is compatible with a specified Dimension'.

hasSomeDimension :: HasDynamicDimension a => a -> Bool

Determines if a value that has a DynamicDimension in fact has any valid dimension at all.