Numeric.Units.Dimensional.Functor

Description

Provides a Functor instance for Dimensional.

Note that this instance is dubious, because it allows you to break the dimensional abstraction. See dmap for more information.

Note that, while this instance overlaps with that given for Dimensionless, it is confluent with that instance.

Note that this is an orphan instance.

Orphan instances

KnownVariant v => Functor (Dimensional v d)
A Functor instance for Dimensional.

Note that this instance is dubious, because it allows you to break the dimensional abstraction. See dmap for more information.

Note that, while this instance overlaps with that given for Dimensionless, it is confluent with that instance.

Note that this is an orphan instance.

Instance details

Methods

fmap :: (a -> b) -> Dimensional v d a -> Dimensional v d b

(<$) :: a -> Dimensional v d b -> Dimensional v d a