Numeric.Units.Dimensional.UnitNames

Description

This module provides types and functions for manipulating unit names.

Please note that the details of the name representation may be less stable than the other APIs provided by this package, as new features using them are still being developed.

Data Types

data UnitName (m :: Metricality)

The name of a unit.

data NameAtom (m :: NameAtomType)

Represents the name of an atomic unit or prefix.

data Prefix

type PrefixName = NameAtom 'PrefixAtom

The name of a metric prefix.

data Metricality

Encodes whether a unit is a metric unit, that is, whether it can be combined with a metric prefix to form a related unit.

Constructors

Metric
Capable of receiving a metric prefix.

NonMetric
Incapable of receiving a metric prefix.

Construction of Unit Names

atom

Arguments

:: String
Interchange name

-> String
Abbreviated name in international English

-> String
Full name in international English

-> UnitName 'NonMetric

Constructs an atomic name for a custom unit.

applyPrefix :: Prefix -> UnitName 'Metric -> UnitName 'NonMetric

Forms a UnitName from a Metric name by applying a metric prefix.

(*) :: UnitName m1 -> UnitName m2 -> UnitName 'NonMetric infixl 7

Form a UnitName by taking the product of two others.

(/) :: UnitName m1 -> UnitName m2 -> UnitName 'NonMetric infixl 7

Form a UnitName by dividing one by another.

(^) :: UnitName m -> Int -> UnitName 'NonMetric infixr 8

Form a UnitName by raising a name to an integer power.

product :: Foldable f => f (UnitName 'NonMetric) -> UnitName 'NonMetric

Forms the product of a list of UnitNames.

If you wish to form a heterogenous product of Metric and NonMetric units you should apply weaken to the Metric ones.

reduce :: UnitName m -> UnitName m

grouped :: UnitName m -> UnitName 'NonMetric

Constructs a UnitName by applying a grouping operation to another UnitName, which may be useful to express precedence.

Standard Names

baseUnitName :: Dimension' -> UnitName 'NonMetric

The name of the base unit associated with a specified dimension.

siPrefixes :: [Prefix]

A list of all Prefixes defined by the SI.

nOne :: UnitName 'NonMetric

The name of the unit of dimensionless values.

Inspecting Prefixes

prefixName :: Prefix -> PrefixName

The name of a metric prefix.

scaleFactor :: Prefix -> Rational

The scale factor denoted by a metric prefix.

Convenience Type Synonyms for Unit Name Transformations

type UnitNameTransformer = forall m. UnitName m -> UnitName 'NonMetric

The type of a unit name transformation that may be associated with an operation that takes a single unit as input.

type UnitNameTransformer2 = forall m1 m2. UnitName m1 -> UnitName m2 -> UnitName 'NonMetric

The type of a unit name transformation that may be associated with an operation that takes two units as input.

Forgetting Unwanted Phantom Types

weaken :: UnitName m -> UnitName 'NonMetric

Convert a UnitName which may or may not be Metric to one which is certainly NonMetric.

strengthen :: UnitName m -> Maybe (UnitName 'Metric)

Attempt to convert a UnitName which may or may not be Metric to one which is certainly Metric.

relax :: forall m1 m2. (Typeable m1, Typeable m2) => UnitName m1 -> Maybe (UnitName m2)

Convert a UnitName of one Metricality into a name of another metricality by strengthening or weakening if neccessary. Because it may not be possible to strengthen, the result is returned in a Maybe wrapper.

name_en :: NameAtom m -> String

The full name of the unit in international English

abbreviation_en :: NameAtom m -> String

The abbreviated name of the unit in international English

asAtomic :: UnitName m -> Maybe (NameAtom ('UnitAtom m))