Numeric.Units.Dimensional.UnitNames.InterchangeNames

Documentation

data InterchangeNameAuthority

Represents the authority which issued an interchange name for a unit.

Constructors

UCUM
The interchange name originated with the Unified Code for Units of Measure.

DimensionalLibrary
The interchange name originated with the dimensional library.

Custom
The interchange name originated with a user of the dimensional library.

data InterchangeName

Constructors

InterchangeName

Fields

name :: String

authority :: InterchangeNameAuthority

isAtomic :: Bool

class HasInterchangeName a where

Determines the authority which issued the interchange name of a unit or unit name. For composite units, this is the least-authoritative interchange name of any constituent name.

Note that the least-authoritative authority is the one sorted as greatest by the Ord instance of InterchangeNameAuthority.

Methods

interchangeName :: a -> InterchangeName