Numeric.Units.Dimensional.Variants

Description

Provides a type level representation of Variants of dimensional values, which may be quantities or units.

Documentation

data Variant

The kind of variants of dimensional values.

Constructors

DQuantity ExactPi'
The value is a quantity, stored as an ExactPi multiple of its value in its dimension's SI coherent unit.

DUnit Metricality
The value is a unit, possibly a Metric unit.

data Metricality

Encodes whether a unit is a metric unit, that is, whether it can be combined with a metric prefix to form a related unit.

Constructors

Metric
Capable of receiving a metric prefix.

NonMetric
Incapable of receiving a metric prefix.

type family (v1 :: Variant) * (v2 :: Variant) :: Variant where ... infixl 7

Forms the product of two Variants.

The product of units is a non-metric unit.

The product of quantities is a quantity.

Equations

('DUnit m1) * ('DUnit m2) = 'DUnit 'NonMetric

('DQuantity s1) * ('DQuantity s2) = 'DQuantity (s1 * s2)

type family (v1 :: Variant) / (v2 :: Variant) :: Variant where ...

Equations

('DUnit m1) / ('DUnit m2) = 'DUnit 'NonMetric

('DQuantity s1) / ('DQuantity s2) = 'DQuantity (s1 / s2)

type family Weaken (v :: Variant) :: Variant where ...

Weakens a Variant by forgetting possibly uninteresting type-level information.

Equations

Weaken ('DQuantity s) = 'DQuantity s

Weaken ('DUnit m) = 'DUnit 'NonMetric

type CompatibleVariants v1 v2 = 'True ~ AreCompatible v1 v2

Two Variants are compatible when dimensional values of the first may be converted into the second merely by changing the representation of their values.